+++ mapchete/__init__.py
@@ -7,7 +7,7 @@ from mapchete._timer import Timer
 
 
 __all__ = ['open', 'count_tiles', 'Mapchete', 'MapcheteProcess', 'Timer']
-__version__ = "0.31"
+__version__ = "0.33"
 
 
 logger = logging.getLogger(__name__)