+++ test/test_cli.py
@@ -7,6 +7,7 @@ import pytest
 from shapely import wkt
 import rasterio
 from rasterio.io import MemoryFile
+from rio_cogeo.cogeo import cog_validate
 import yaml
 
 import mapchete
@@ -19,7 +20,6 @@ def run_cli(args, expected_exit_code=0, output_contains=None, raise_exc=True):
     if output_contains:
         assert output_contains in result.output
     if raise_exc and result.exception:
-        print(result.output)
         raise result.exception
     assert result.exit_code == expected_exit_code
 
@@ -52,7 +52,7 @@ def test_missing_input_file(mp_tmpdir):
     run_cli(
         ["execute", "process.mapchete", "--input-file", "invalid.tif"],
         expected_exit_code=2,
-        output_contains='Path "process.mapchete" does not exist.',
+        output_contains="Path 'process.mapchete' does not exist.",
         raise_exc=False
     )
 
@@ -65,7 +65,7 @@ def test_create_and_execute(mp_tmpdir, cleantopo_br_tif):
     # create from template
     run_cli(
         [
-            'create', temp_mapchete, temp_process, out_format,
+            "create", temp_mapchete, temp_process, out_format,
             "--pyramid-type", "geodetic"
         ],
         expected_exit_code=0
@@ -80,8 +80,8 @@ def test_create_and_execute(mp_tmpdir, cleantopo_br_tif):
     with pytest.raises(MapcheteProcessOutputError):
         run_cli(
             [
-                'execute', temp_mapchete, '--tile', '6', '62', '124',
-                '--input-file', cleantopo_br_tif, '-d'
+                "execute", temp_mapchete, "--tile", "6", "62", "124",
+                "--input-file", cleantopo_br_tif, "-d"
             ],
             expected_exit_code=-1
         )
@@ -94,7 +94,7 @@ def test_create_existing(mp_tmpdir):
     out_format = "GTiff"
     # create files from template
     args = [
-        'create', temp_mapchete, temp_process, out_format, "--pyramid-type", "geodetic"
+        "create", temp_mapchete, temp_process, out_format, "--pyramid-type", "geodetic"
     ]
     run_cli(args)
     # try to create again
@@ -109,7 +109,7 @@ def test_execute_multiprocessing(mp_tmpdir, cleantopo_br, cleantopo_br_tif):
     out_format = "GTiff"
     # create from template
     run_cli([
-        'create',
+        "create",
         temp_mapchete,
         temp_process,
         out_format,
@@ -124,20 +124,20 @@ def test_execute_multiprocessing(mp_tmpdir, cleantopo_br, cleantopo_br_tif):
     # run process with multiprocessing
     with pytest.raises(MapcheteProcessOutputError):
         run_cli([
-            'execute',
+            "execute",
             temp_mapchete,
-            '--zoom', '5',
-            '--input-file', cleantopo_br_tif,
-            '-m', '2',
-            '-d'
+            "--zoom", "5",
+            "--input-file", cleantopo_br_tif,
+            "-m", "2",
+            "-d"
         ])
     # run example process with multiprocessing
     run_cli([
-        'execute',
+        "execute",
         cleantopo_br.path,
-        '--zoom', '5',
-        '-m', '2',
-        '-d'
+        "--zoom", "5",
+        "-m", "2",
+        "-d"
     ])
 
 
@@ -145,7 +145,7 @@ def test_execute_debug(mp_tmpdir, example_mapchete):
     """Using debug output."""
     run_cli(
         [
-            'execute', example_mapchete.path, "-t", "10", "500", "1040",
+            "execute", example_mapchete.path, "-t", "10", "500", "1040",
             "--debug"
         ]
     )
@@ -153,7 +153,7 @@ def test_execute_debug(mp_tmpdir, example_mapchete):
 
 def test_execute_vrt(mp_tmpdir, cleantopo_br):
     """Using debug output."""
-    run_cli(['execute', cleantopo_br.path, "-z", "5", "--vrt"])
+    run_cli(["execute", cleantopo_br.path, "-z", "5", "--vrt"])
     with mapchete.open(cleantopo_br.dict) as mp:
         vrt_path = os.path.join(mp.config.output.path, "5.vrt")
         with rasterio.open(vrt_path) as src:
@@ -161,7 +161,7 @@ def test_execute_vrt(mp_tmpdir, cleantopo_br):
 
     # run again, this time with custom output directory
     run_cli(
-        ['execute', cleantopo_br.path, "-z", "5", "--vrt", "--idx-out-dir", mp_tmpdir]
+        ["execute", cleantopo_br.path, "-z", "5", "--vrt", "--idx-out-dir", mp_tmpdir]
     )
     with mapchete.open(cleantopo_br.dict) as mp:
         vrt_path = os.path.join(mp_tmpdir, "5.vrt")
@@ -170,25 +170,25 @@ def test_execute_vrt(mp_tmpdir, cleantopo_br):
 
     # run with single tile
     run_cli(
-        ['execute', cleantopo_br.path, "-t", "5", "3", "7", "--vrt"]
+        ["execute", cleantopo_br.path, "-t", "5", "3", "7", "--vrt"]
     )
 
     # no new entries
     run_cli(
-        ['execute', cleantopo_br.path, "-t", "5", "0", "0", "--vrt"]
+        ["execute", cleantopo_br.path, "-t", "5", "0", "0", "--vrt"]
     )
 
 
 def test_execute_verbose(mp_tmpdir, example_mapchete):
     """Using verbose output."""
-    run_cli(['execute', example_mapchete.path, "-t", "10", "500", "1040", "--verbose"])
+    run_cli(["execute", example_mapchete.path, "-t", "10", "500", "1040", "--verbose"])
 
 
 def test_execute_logfile(mp_tmpdir, example_mapchete):
     """Using logfile."""
     logfile = os.path.join(mp_tmpdir, "temp.log")
     run_cli([
-        'execute', example_mapchete.path, "-t", "10", "500", "1040", "--logfile", logfile
+        "execute", example_mapchete.path, "-t", "10", "500", "1040", "--logfile", logfile
     ])
     assert os.path.isfile(logfile)
     with open(logfile) as log:
@@ -197,34 +197,34 @@ def test_execute_logfile(mp_tmpdir, example_mapchete):
 
 def test_execute_wkt_bounds(mp_tmpdir, example_mapchete, wkt_geom):
     """Using bounds from WKT."""
-    run_cli(['execute', example_mapchete.path, "--wkt-geometry", wkt_geom])
+    run_cli(["execute", example_mapchete.path, "--wkt-geometry", wkt_geom])
 
 
 def test_execute_point(mp_tmpdir, example_mapchete, wkt_geom):
     """Using bounds from WKT."""
     g = wkt.loads(wkt_geom)
     run_cli([
-        'execute', example_mapchete.path,
+        "execute", example_mapchete.path,
         "--point", str(g.centroid.x), str(g.centroid.y)
     ])
 
 
 def test_formats(capfd):
     """Output of mapchete formats command."""
-    run_cli(['formats'])
+    run_cli(["formats"])
     err = capfd.readouterr()[1]
     assert not err
-    run_cli(['formats', '-i'])
+    run_cli(["formats", "-i"])
     err = capfd.readouterr()[1]
     assert not err
-    run_cli(['formats', '-o'])
+    run_cli(["formats", "-o"])
     err = capfd.readouterr()[1]
     assert not err
 
 
 def test_convert_geodetic(cleantopo_br_tif, mp_tmpdir):
     """Automatic geodetic tile pyramid creation of raster files."""
-    run_cli(['convert', cleantopo_br_tif, mp_tmpdir, "--output-pyramid", "geodetic"])
+    run_cli(["convert", cleantopo_br_tif, mp_tmpdir, "--output-pyramid", "geodetic"])
     for zoom, row, col in [(4, 15, 31), (3, 7, 15), (2, 3, 7), (1, 1, 3)]:
         out_file = os.path.join(
             *[mp_tmpdir, str(zoom), str(row), str(col) + ".tif"])
@@ -237,7 +237,7 @@ def test_convert_geodetic(cleantopo_br_tif, mp_tmpdir):
 
 def test_convert_mercator(cleantopo_br_tif, mp_tmpdir):
     """Automatic mercator tile pyramid creation of raster files."""
-    run_cli(['convert', cleantopo_br_tif, mp_tmpdir, "--output-pyramid", "mercator"])
+    run_cli(["convert", cleantopo_br_tif, mp_tmpdir, "--output-pyramid", "mercator"])
     for zoom, row, col in [(4, 15, 15), (3, 7, 7)]:
         out_file = os.path.join(*[mp_tmpdir, str(zoom), str(row), str(col) + ".tif"])
         with rasterio.open(out_file, "r") as src:
@@ -250,7 +250,7 @@ def test_convert_mercator(cleantopo_br_tif, mp_tmpdir):
 def test_convert_png(cleantopo_br_tif, mp_tmpdir):
     """Automatic PNG tile pyramid creation of raster files."""
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         mp_tmpdir,
         "--output-pyramid", "mercator",
@@ -266,11 +266,30 @@ def test_convert_png(cleantopo_br_tif, mp_tmpdir):
             assert data.mask.any()
 
 
+def test_convert_bidx(cleantopo_br_tif, mp_tmpdir):
+    """Automatic geodetic tile pyramid creation of raster files."""
+    single_gtiff = os.path.join(mp_tmpdir, "single_out_bidx.tif")
+    run_cli([
+        "convert",
+        cleantopo_br_tif,
+        single_gtiff,
+        "--output-pyramid", "geodetic",
+        "-z", "3",
+        "--bidx", "1"
+    ])
+    with rasterio.open(single_gtiff, "r") as src:
+        assert src.meta["driver"] == "GTiff"
+        assert src.meta["dtype"] == "uint16"
+        data = src.read(masked=True)
+        assert data.mask.any()
+        assert not src.overviews(1)
+
+
 def test_convert_single_gtiff(cleantopo_br_tif, mp_tmpdir):
     """Automatic geodetic tile pyramid creation of raster files."""
     single_gtiff = os.path.join(mp_tmpdir, "single_out.tif")
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         single_gtiff,
         "--output-pyramid", "geodetic",
@@ -284,11 +303,31 @@ def test_convert_single_gtiff(cleantopo_br_tif, mp_tmpdir):
         assert not src.overviews(1)
 
 
+def test_convert_single_gtiff_cog(cleantopo_br_tif, mp_tmpdir):
+    """Automatic geodetic tile pyramid creation of raster files."""
+    single_gtiff = os.path.join(mp_tmpdir, "single_out_cog.tif")
+    run_cli([
+        "convert",
+        cleantopo_br_tif,
+        single_gtiff,
+        "--output-pyramid", "geodetic",
+        "-z", "3",
+        "--cog"
+    ])
+    with rasterio.open(single_gtiff, "r") as src:
+        assert src.meta["driver"] == "GTiff"
+        assert src.meta["dtype"] == "uint16"
+        data = src.read(masked=True)
+        assert data.mask.any()
+    assert cog_validate(single_gtiff, strict=True)
+
+
+
 def test_convert_single_gtiff_overviews(cleantopo_br_tif, mp_tmpdir):
     """Automatic geodetic tile pyramid creation of raster files."""
     single_gtiff = os.path.join(mp_tmpdir, "single_out.tif")
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         single_gtiff,
         "--output-pyramid", "geodetic",
@@ -308,7 +347,7 @@ def test_convert_remote_single_gtiff(http_raster, mp_tmpdir):
     """Automatic geodetic tile pyramid creation of raster files."""
     single_gtiff = os.path.join(mp_tmpdir, "single_out.tif")
     run_cli([
-        'convert',
+        "convert",
         http_raster,
         single_gtiff,
         "--output-pyramid", "geodetic",
@@ -324,7 +363,7 @@ def test_convert_remote_single_gtiff(http_raster, mp_tmpdir):
 def test_convert_dtype(cleantopo_br_tif, mp_tmpdir):
     """Automatic tile pyramid creation using dtype scale."""
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         mp_tmpdir,
         "--output-pyramid", "mercator",
@@ -342,7 +381,7 @@ def test_convert_dtype(cleantopo_br_tif, mp_tmpdir):
 def test_convert_scale_ratio(cleantopo_br_tif, mp_tmpdir):
     """Automatic tile pyramid creation cropping data."""
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         mp_tmpdir,
         "--output-pyramid", "mercator",
@@ -362,7 +401,7 @@ def test_convert_scale_ratio(cleantopo_br_tif, mp_tmpdir):
 def test_convert_scale_offset(cleantopo_br_tif, mp_tmpdir):
     """Automatic tile pyramid creation cropping data."""
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         mp_tmpdir,
         "--output-pyramid", "mercator",
@@ -383,7 +422,7 @@ def test_convert_clip(cleantopo_br_tif, mp_tmpdir, landpoly):
     """Automatic tile pyramid creation cropping data."""
     run_cli(
         [
-            'convert',
+            "convert",
             cleantopo_br_tif,
             mp_tmpdir,
             "--output-pyramid", "geodetic",
@@ -397,7 +436,7 @@ def test_convert_clip(cleantopo_br_tif, mp_tmpdir, landpoly):
 def test_convert_zoom(cleantopo_br_tif, mp_tmpdir):
     """Automatic tile pyramid creation using a specific zoom."""
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         mp_tmpdir,
         "--output-pyramid", "mercator",
@@ -411,7 +450,7 @@ def test_convert_zoom(cleantopo_br_tif, mp_tmpdir):
 def test_convert_zoom_minmax(cleantopo_br_tif, mp_tmpdir):
     """Automatic tile pyramid creation using min max zoom."""
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         mp_tmpdir,
         "--output-pyramid", "mercator",
@@ -425,7 +464,7 @@ def test_convert_zoom_minmax(cleantopo_br_tif, mp_tmpdir):
 def test_convert_zoom_maxmin(cleantopo_br_tif, mp_tmpdir):
     """Automatic tile pyramid creation using max min zoom."""
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br_tif,
         mp_tmpdir,
         "--output-pyramid", "mercator",
@@ -442,7 +481,7 @@ def test_convert_mapchete(cleantopo_br, mp_tmpdir):
     with mapchete.open(cleantopo_br.path) as mp:
         mp.batch_process(zoom=[1, 4])
     run_cli([
-        'convert',
+        "convert",
         cleantopo_br.path,
         mp_tmpdir,
         "--output-pyramid", "geodetic",
@@ -463,7 +502,7 @@ def test_convert_tiledir(cleantopo_br, mp_tmpdir):
     with mapchete.open(cleantopo_br.path) as mp:
         mp.batch_process(zoom=[1, 4])
     run_cli([
-        'convert',
+        "convert",
         os.path.join(
             cleantopo_br.dict["config_dir"], cleantopo_br.dict["output"]["path"]
         ),
@@ -509,7 +548,7 @@ def test_convert_errors(s2_band_jp2, mp_tmpdir, s2_band, cleantopo_br, landpoly)
     # zoom level required
     run_cli(
         [
-            'convert',
+            "convert",
             tiledir_path,
             mp_tmpdir,
             "--output-pyramid", "geodetic",
@@ -522,7 +561,7 @@ def test_convert_errors(s2_band_jp2, mp_tmpdir, s2_band, cleantopo_br, landpoly)
     # incompatible formats
     run_cli(
         [
-            'convert',
+            "convert",
             tiledir_path,
             mp_tmpdir,
             "--output-pyramid", "geodetic",
@@ -539,7 +578,7 @@ def test_convert_errors(s2_band_jp2, mp_tmpdir, s2_band, cleantopo_br, landpoly)
     # unsupported format
     run_cli(
         [
-            'convert',
+            "convert",
             landpoly,
             mp_tmpdir,
             "--output-pyramid", "geodetic",
@@ -554,9 +593,9 @@ def test_convert_errors(s2_band_jp2, mp_tmpdir, s2_band, cleantopo_br, landpoly)
     # unsupported output format extension
     run_cli(
         [
-            'convert',
+            "convert",
             s2_band_jp2,
-            'output.jp2',
+            "output.jp2",
             "--output-pyramid", "geodetic",
             "--zoom", "5",
         ],
@@ -565,30 +604,43 @@ def test_convert_errors(s2_band_jp2, mp_tmpdir, s2_band, cleantopo_br, landpoly)
         raise_exc=False
     )
 
+    # malformed band index
+    run_cli(
+        [
+            "convert",
+            s2_band_jp2,
+            "output.tif",
+            "--bidx", "invalid"
+        ],
+        expected_exit_code=2,
+        output_contains=("Invalid value for '--bidx'"),
+        raise_exc=False
+    )
+
 
 def test_serve_cli_params(cleantopo_br, mp_tmpdir):
     """Test whether different CLI params pass."""
     # assert too few arguments error
     with pytest.raises(SystemExit):
-        run_cli(['serve'])
+        run_cli(["serve"])
 
     for args in [
-        ['serve', cleantopo_br.path],
-        ['serve', cleantopo_br.path, "--port", "5001"],
-        ['serve', cleantopo_br.path, "--internal-cache", "512"],
-        ['serve', cleantopo_br.path, "--zoom", "5"],
-        ['serve', cleantopo_br.path, "--bounds", "-1", "-1", "1", "1"],
-        ['serve', cleantopo_br.path, "--overwrite"],
-        ['serve', cleantopo_br.path, "--readonly"],
-        ['serve', cleantopo_br.path, "--memory"],
-        ['serve', cleantopo_br.path, "--input-file", cleantopo_br.path],
+        ["serve", cleantopo_br.path],
+        ["serve", cleantopo_br.path, "--port", "5001"],
+        ["serve", cleantopo_br.path, "--internal-cache", "512"],
+        ["serve", cleantopo_br.path, "--zoom", "5"],
+        ["serve", cleantopo_br.path, "--bounds", "-1", "-1", "1", "1"],
+        ["serve", cleantopo_br.path, "--overwrite"],
+        ["serve", cleantopo_br.path, "--readonly"],
+        ["serve", cleantopo_br.path, "--memory"],
+        ["serve", cleantopo_br.path, "--input-file", cleantopo_br.path],
     ]:
         run_cli(args)
 
 
 def test_serve(client, mp_tmpdir):
     """Mapchete serve with default settings."""
-    tile_base_url = '/wmts_simple/1.0.0/dem_to_hillshade/default/WGS84/'
+    tile_base_url = "/wmts_simple/1.0.0/dem_to_hillshade/default/WGS84/"
     for url in ["/"]:
         response = client.get(url)
         assert response.status_code == 200
@@ -621,10 +673,10 @@ def test_serve(client, mp_tmpdir):
 
 def test_index_geojson(mp_tmpdir, cleantopo_br):
     # execute process at zoom 3
-    run_cli(['execute', cleantopo_br.path, '-z', '3', '--debug'])
+    run_cli(["execute", cleantopo_br.path, "-z", "3", "--debug"])
 
     # generate index for zoom 3
-    run_cli(['index', cleantopo_br.path,  '-z', '3', '--geojson', '--debug'])
+    run_cli(["index", cleantopo_br.path, "-z", "3", "--geojson", "--debug"])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert len(files) == 3
@@ -637,12 +689,12 @@ def test_index_geojson(mp_tmpdir, cleantopo_br):
 
 def test_index_geojson_fieldname(mp_tmpdir, cleantopo_br):
     # execute process at zoom 3
-    run_cli(['execute', cleantopo_br.path, '-z', '3', '--debug'])
+    run_cli(["execute", cleantopo_br.path, "-z", "3", "--debug"])
 
     # index and rename "location" to "new_fieldname"
     run_cli([
-        'index', cleantopo_br.path,  '-z', '3', '--geojson', '--debug',
-        '--fieldname', 'new_fieldname'
+        "index", cleantopo_br.path, "-z", "3", "--geojson", "--debug",
+        "--fieldname", "new_fieldname"
     ])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
@@ -655,13 +707,13 @@ def test_index_geojson_fieldname(mp_tmpdir, cleantopo_br):
 
 def test_index_geojson_basepath(mp_tmpdir, cleantopo_br):
     # execute process at zoom 3
-    run_cli(['execute', cleantopo_br.path, '-z', '3', '--debug'])
+    run_cli(["execute", cleantopo_br.path, "-z", "3", "--debug"])
 
-    basepath = 'http://localhost'
+    basepath = "http://localhost"
     # index and rename "location" to "new_fieldname"
     run_cli([
-        'index', cleantopo_br.path,  '-z', '3', '--geojson', '--debug',
-        '--basepath', basepath
+        "index", cleantopo_br.path, "-z", "3", "--geojson", "--debug",
+        "--basepath", basepath
     ])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
@@ -674,13 +726,13 @@ def test_index_geojson_basepath(mp_tmpdir, cleantopo_br):
 
 def test_index_geojson_for_gdal(mp_tmpdir, cleantopo_br):
     # execute process at zoom 3
-    run_cli(['execute', cleantopo_br.path, '-z', '3', '--debug'])
+    run_cli(["execute", cleantopo_br.path, "-z", "3", "--debug"])
 
-    basepath = 'http://localhost'
+    basepath = "http://localhost"
     # index and rename "location" to "new_fieldname"
     run_cli([
-        'index', cleantopo_br.path,  '-z', '3', '--geojson', '--debug',
-        '--basepath', basepath, '--for-gdal'
+        "index", cleantopo_br.path, "-z", "3", "--geojson", "--debug",
+        "--basepath", basepath, "--for-gdal"
     ])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
@@ -694,9 +746,9 @@ def test_index_geojson_for_gdal(mp_tmpdir, cleantopo_br):
 
 def test_index_geojson_tile(mp_tmpdir, cleantopo_tl):
     # execute process for single tile
-    run_cli(['execute', cleantopo_tl.path, '-t', '3', '0', '0', '--debug'])
+    run_cli(["execute", cleantopo_tl.path, "-t", "3", "0", "0", "--debug"])
     # generate index
-    run_cli(['index', cleantopo_tl.path, '-t', '3', '0', '0', '--geojson', '--debug'])
+    run_cli(["index", cleantopo_tl.path, "-t", "3", "0", "0", "--geojson", "--debug"])
     with mapchete.open(cleantopo_tl.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert len(files) == 3
@@ -707,11 +759,11 @@ def test_index_geojson_tile(mp_tmpdir, cleantopo_tl):
 
 def test_index_geojson_wkt_geom(mp_tmpdir, cleantopo_br, wkt_geom):
     # execute process at zoom 3
-    run_cli(['execute', cleantopo_br.path, '--debug', "--wkt-geometry", wkt_geom])
+    run_cli(["execute", cleantopo_br.path, "--debug", "--wkt-geometry", wkt_geom])
 
     # generate index for zoom 3
     run_cli([
-        'index', cleantopo_br.path, '--geojson', '--debug', "--wkt-geometry", wkt_geom
+        "index", cleantopo_br.path, "--geojson", "--debug", "--wkt-geometry", wkt_geom
     ])
 
     with mapchete.open(cleantopo_br.dict) as mp:
@@ -722,10 +774,10 @@ def test_index_geojson_wkt_geom(mp_tmpdir, cleantopo_br, wkt_geom):
 
 def test_index_gpkg(mp_tmpdir, cleantopo_br):
     # execute process
-    run_cli(['execute', cleantopo_br.path, '-z', '5', '--debug'])
+    run_cli(["execute", cleantopo_br.path, "-z", "5", "--debug"])
 
     # generate index
-    run_cli(['index', cleantopo_br.path,  '-z', '5', '--gpkg', '--debug'])
+    run_cli(["index", cleantopo_br.path, "-z", "5", "--gpkg", "--debug"])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert "5.gpkg" in files
@@ -735,7 +787,7 @@ def test_index_gpkg(mp_tmpdir, cleantopo_br):
         assert len(list(src)) == 1
 
     # write again and assert there is no new entry because there is already one
-    run_cli(['index', cleantopo_br.path,  '-z', '5', '--gpkg', '--debug'])
+    run_cli(["index", cleantopo_br.path, "-z", "5", "--gpkg", "--debug"])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert "5.gpkg" in files
@@ -747,10 +799,10 @@ def test_index_gpkg(mp_tmpdir, cleantopo_br):
 
 def test_index_shp(mp_tmpdir, cleantopo_br):
     # execute process
-    run_cli(['execute', cleantopo_br.path, '-z', '5', '--debug'])
+    run_cli(["execute", cleantopo_br.path, "-z", "5", "--debug"])
 
     # generate index
-    run_cli(['index', cleantopo_br.path,  '-z', '5', '--shp', '--debug'])
+    run_cli(["index", cleantopo_br.path, "-z", "5", "--shp", "--debug"])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert "5.shp" in files
@@ -760,7 +812,7 @@ def test_index_shp(mp_tmpdir, cleantopo_br):
         assert len(list(src)) == 1
 
     # write again and assert there is no new entry because there is already one
-    run_cli(['index', cleantopo_br.path,  '-z', '5', '--shp', '--debug'])
+    run_cli(["index", cleantopo_br.path, "-z", "5", "--shp", "--debug"])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert "5.shp" in files
@@ -772,10 +824,10 @@ def test_index_shp(mp_tmpdir, cleantopo_br):
 
 def test_index_text(cleantopo_br):
     # execute process
-    run_cli(['execute', cleantopo_br.path, '-z', '5', '--debug'])
+    run_cli(["execute", cleantopo_br.path, "-z", "5", "--debug"])
 
     # generate index
-    run_cli(['index', cleantopo_br.path,  '-z', '5', '--txt', '--debug'])
+    run_cli(["index", cleantopo_br.path, "-z", "5", "--txt", "--debug"])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert "5.txt" in files
@@ -786,7 +838,7 @@ def test_index_text(cleantopo_br):
             assert l.endswith("7.tif\n")
 
     # write again and assert there is no new entry because there is already one
-    run_cli(['index', cleantopo_br.path,  '-z', '5', '--txt', '--debug'])
+    run_cli(["index", cleantopo_br.path, "-z", "5", "--txt", "--debug"])
     with mapchete.open(cleantopo_br.dict) as mp:
         files = os.listdir(mp.config.output.path)
         assert "5.txt" in files
@@ -799,23 +851,23 @@ def test_index_text(cleantopo_br):
 
 def test_index_errors(mp_tmpdir, cleantopo_br):
     with pytest.raises(SystemExit):
-        run_cli(['index', cleantopo_br.path,  '-z', '5', '--debug'])
+        run_cli(["index", cleantopo_br.path, "-z", "5", "--debug"])
 
 
 def test_processes():
-    run_cli(['processes'])
-    run_cli(['processes', '-n', 'mapchete.processes.examples.example_process'])
+    run_cli(["processes"])
+    run_cli(["processes", "-n", "mapchete.processes.examples.example_process"])
 
 
 def test_callback_errors(cleantopo_tl):
     run_cli(
-        ['execute', cleantopo_tl.path, '--zoom', '4,5,7'],
+        ["execute", cleantopo_tl.path, "--zoom", "4,5,7"],
         expected_exit_code=2,
         raise_exc=False,
         output_contains="zooms can be maximum two items"
     )
     run_cli(
-        ['execute', cleantopo_tl.path, '--zoom', 'invalid'],
+        ["execute", cleantopo_tl.path, "--zoom", "invalid"],
         expected_exit_code=2,
         raise_exc=False,
         output_contains="zoom levels must be integer values"